% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletSurvey.R
\name{leafletSurvey}
\alias{leafletSurvey}
\title{leafletSurvey}
\usage{
leafletSurvey(
  trackdata = NULL,
  obsdata = NULL,
  boundary = NULL,
  footprint = NULL,
  trackColour = "transect",
  obsColour = "behaviour"
)
}
\arguments{
\item{trackdata}{An sf dataframe containing track points or lines}

\item{obsdata}{An sf dataframe containing observation points.}

\item{boundary}{An optional sf polygon representing the survey boundary}

\item{obsColour}{Character; column name in obsdata to colour points by. Options are "behaviour", "species", or "none" for no colouring.}
}
\value{
A leaflet map object displaying the tracks.
}
\description{
Plot survey data using Leaflet
}
\details{
This function creates an interactive map using the leaflet package to visualize survey tracks and observations.


This function is designed to help visualize spatial track data interactively. It can be customized with additional leaflet layers and options.
}
\examples{
\dontrun{
plotTracks.leaflet(track_data)
}

}
