% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHiDef.R
\name{readHiDef}
\alias{readHiDef}
\title{readHiDef}
\usage{
readHiDef(filepath, overwrite_dates = FALSE)
}
\arguments{
\item{filepath}{The filepath containing relevant HiDef datafiles.}

\item{overwrite_dates}{Logical; if TRUE, dates in the data will be replaced with dates extracted from the filepaths. Default is FALSE.}
}
\value{
A list containing two sf dataframes: 'tracks' and 'observations'.
}
\description{
Calls HiDef data from a defined filepath and compiles data into track and
observation objects.
}
\details{
TODO:
- Resolve timestamp formatting in both track and observations
}
